# perf: Arena based ast