--- a/crates/swc_allocator/src/arena/alloc.rs
+++ b/crates/swc_allocator/src/arena/alloc.rs
@@ -1,0 +1,23 @@
+use std::ops::{Deref, DerefMut};
+
+use bumpalo::Bump;
+
+/// The actual storage for [FastAlloc].
+#[derive(Default)]
+pub struct Allocator {
+    alloc: Bump,
+}
+
+impl Deref for Allocator {
+    type Target = Bump;
+
+    fn deref(&self) -> &Bump {
+        &self.alloc
+    }
+}
+
+impl DerefMut for Allocator {
+    fn deref_mut(&mut self) -> &mut Bump {
+        &mut self.alloc
+    }
+}