--- a/crates/swc_allocator/src/arena/alloc.rs
+++ b/crates/swc_allocator/src/arena/alloc.rs
@@ -3,7 +3,7 @@ use std::ops::{Deref, DerefMut};
 use bumpalo::Bump;
 
 /// The actual storage for [FastAlloc].
-#[derive(Default)]
+#[derive(Debug, Default)]
 pub struct Allocator {
     alloc: Bump,
 }