--- a/crates/swc_allocator/src/arena/mod.rs
+++ b/crates/swc_allocator/src/arena/mod.rs
@@ -1,0 +1,12 @@
+//! Experimental arena allocator
+
+mod alloc;
+mod boxed;
+mod clone_in;
+mod vec;
+
+pub use alloc::Allocator;
+
+pub use boxed::Box;
+pub use clone_in::CloneIn;
+pub use vec::Vec;