--- a/crates/swc_ecma_parser/src/arena/mod.rs
+++ b/crates/swc_ecma_parser/src/arena/mod.rs
@@ -1,0 +1,1 @@
+pub mod parser;