… Cargo.lock
… crates/ast_node_arena/Cargo.toml
… crates/swc_allocator/Cargo.toml
… crates/swc_allocator/src/arena/vec/mod.rs
… crates/swc_allocator/src/arena/vec/rkyv.rs
… crates/swc_allocator/src/arena/vec/serde.rs
… crates/swc_allocator/src/lib.rs
… crates/swc_atoms/Cargo.toml
… crates/swc_atoms/src/lib.rs
… crates/swc_common/Cargo.toml
… crates/swc_common/src/arena.rs
… crates/swc_common/src/lib.rs
… crates/swc_common/src/syntax_pos.rs
… crates/swc_ecma_ast/Cargo.toml
… crates/swc_ecma_ast/src/arena/class.rs
… crates/swc_ecma_ast/src/arena/decl.rs
… crates/swc_ecma_ast/src/arena/expr.rs
… crates/swc_ecma_ast/src/arena/function.rs
… crates/swc_ecma_ast/src/arena/ident.rs
… crates/swc_ecma_ast/src/arena/jsx.rs
… crates/swc_ecma_ast/src/arena/lit.rs
… crates/swc_ecma_ast/src/arena/mod.rs
… crates/swc_ecma_ast/src/arena/module.rs
… crates/swc_ecma_ast/src/arena/module_decl.rs
… crates/swc_ecma_ast/src/arena/pat.rs
… crates/swc_ecma_ast/src/arena/prop.rs
… crates/swc_ecma_ast/src/arena/stmt.rs
… crates/swc_ecma_ast/src/arena/typescript.rs
… crates/swc_ecma_ast/src/operators.rs
… crates/swc_ecma_parser/Cargo.toml