# How to do it
The process of refactoring is very mechanical and labor intensive:
1. Declare lifetimes for all ast structs and other related structs and traits.
2. Create all ast nodes using bumpalo allocator.
3. Fix lots of compilation errors.

# Effectiveness
I have almost finished `swc_ecma_ast`, `swc_ecma_parser` and `swc_ecma_visit` crates. You can run the benchmark with `cargo bench -p swc_ecma_parser --bench arena_compare`:

|   | Parser | VisitMut |
| -- | -- | -- |
| Swc baseline | 12.963ms | 424.98us |
| Swc arena     |  11.881ms | 374.05us |
| Oxc | 4.3084ms | 372.51us |

# Existing Barriers

When I was refactoring the parser, I found the following steps to be very mechanical and cumbersome. I think it will greatly reduce the developer enthusiasm for refactoring codebase and downstream applications. 

1. **Unavailable**: Ast conversion by bridging `Into`.
2. **Unavailable**: Pattern match with struct destruction 
3. **Unavailable**: `AstNode::default()`
4. **Unavailable**: Direct deserialization.
5. Wrap many ast constructions in `Box::new_in(.., self.alloc)`